added student table component
