+++ log-o-matic/src/router/index.ts
@@ -5,6 +5,7 @@ import Home from "../views/Home.vue";
 import Admin from "../views/Admin.vue";
 import ShowLogs from "../views/ShowLogs.vue";
 import Login from "../views/Login.vue";
+import Student from "../views/Student.vue";
 
 Vue.use(VueRouter);
 
@@ -30,7 +31,7 @@ const routes: Array<RouteConfig> = [
     name: "Logs",
     component: ShowLogs,
     meta: {
-      allowedRoles: ["admin", "teacher", "student"]
+      allowedRoles: ["admin", "teacher"]
     }
   },
   {
@@ -39,6 +40,14 @@ const routes: Array<RouteConfig> = [
     component: Admin,
     meta: {
       allowedRoles: ["admin"]
+    }
+  },
+  {
+    path: "/student",
+    name: "Student",
+    component: Student,
+    meta: {
+      allowedRoles: ["student"]
     }
   }
 ];
@@ -57,6 +66,8 @@ router.beforeEach((to, from, next) => {
   } else {
     if (userRole === "guest") {
       next("/");
+    } else if (userRole === "student") {
+      next("/student");
     } else {
       next("/logs");
     }
