+++ log-o-matic/src/views/showLogs.vue
@@ -4,7 +4,7 @@
 
 <script lang="ts">
 import Vue from "vue";
-import ShowLogs from "@/components/ShowLog.vue";
+import ShowLogs from "@/components/ShowLogs.vue";
 export default Vue.extend({
   name: "ViewLogs",
   components: { ShowLogs }
